requirements.txt
dotenv
python-dotenv
pydantic
httpx
pillow
langgraph
openai
chromadb
aiofiles
mss
numpy
rich
uvicorn
fastapi